backend/requirements.txt
fastapi==0.135.2
uvicorn[standard]==0.42.0
python-socketio==5.16.1
sqlalchemy==2.0.48
asyncpg==0.31.0
alembic==1.18.4
celery==5.6.3
redis==7.4.0
aioredis==2.0.1
pydantic==2.12.5
pydantic-settings==2.13.1
python-dotenv==1.2.2
PyJWT==2.12.1
argon2-cffi==25.1.0
python-multipart==0.0.22
pwdlib[argon2]==0.3.0
aiosqlite==0.22.1